separated socket.io logic from the index file
